Create styles.css
This file was created by Laura Kalbag for the Perch sample site. To the extent these items are found to not be used, they can be eliminated, with the goal of covering styles in the specific Willamette towers css files. Once all these files have been subsumed into the WT css files, this file can be eliminated.