Re-order minutes regions from original
Re-orders the minutes regions so that the file upload appears first. Office staff may do only this action to update minutes, and having it first is for their convenience. Other regions offer optional text to appear on the page and the URL for minutes configured to appear as a web page.